requirements.txt
openai
pandas
numpy
jupyter
pyyaml
transformers
beautifulsoup4>=4.12.0
lxml>=4.9.0